Create a Interface for Scheduler Impl and following up changes
